+++ src/eslint.js
@@ -12,7 +12,7 @@ function buildConfig(parameters, globals) {
   if (globals && globals.sourceFiles) {
     defaults.files = globals.sourceFiles;
   } else if (globals) {
-    globals.sourceFiles = defaults.files;
+    globals.sourceFiles = defaults.files; // eslint-disable-line no-param-reassign
   }
 
   return configHelper.mergeParameters(defaults, parameters);
@@ -53,7 +53,7 @@ module.exports = {
    * );
    */
   register(gulp, taskName, parameters, globals) {
-    const installDependenciesTaskName = taskName + '-install-dependencies';
+    const installDependenciesTaskName = `${taskName}-install-dependencies`;
     npmInstall.register(gulp, installDependenciesTaskName, {
       dependencies: this.dependencies
     });
@@ -66,11 +66,13 @@ module.exports = {
         .pipe(eslint.failAfterError());
     };
 
-    gulp.task(taskName, [installDependenciesTaskName], () => {
-      return validate(buildConfig(parameters, globals));
-    });
+    gulp.task(
+      taskName,
+      [installDependenciesTaskName],
+      () => validate(buildConfig(parameters, globals))
+    );
 
-    gulp.task(taskName + ':fix', [installDependenciesTaskName], (done) => {
+    gulp.task(`${taskName}:fix`, [installDependenciesTaskName], (done) => {
       const gulpIf = require('gulp-if');
       const filesByFolderPath = {};
       const config = buildConfig(parameters, globals);
@@ -84,8 +86,8 @@ module.exports = {
         })
         .on('end', () => {
           Promise.all(
-            Object.keys(filesByFolderPath).map(folderPath => {
-              return new Promise((resolve, reject) => {
+            Object.keys(filesByFolderPath)
+              .map(folderPath => new Promise((resolve, reject) => {
                 validate(Object.assign({}, config, {
                   files: filesByFolderPath[folderPath],
                   fix: true
@@ -98,8 +100,7 @@ module.exports = {
                   .on('end', () => resolve())
                   .on('data', () => {
                   });
-              });
-            })
+              }))
           ).then(() => done(), err => done(err));
         });
     });