+++ src/jest.js
@@ -11,7 +11,7 @@ function buildConfig(parameters, globals) {
   const defaults = require('./jest-defaults');
 
   if (globals && !globals.babel) {
-    globals.babel = require('./lib/global-babel');
+    globals.babel = require('./lib/global-babel'); // eslint-disable-line no-param-reassign
   }
 
   return configHelper.mergeParameters(defaults, parameters);
@@ -51,12 +51,12 @@ module.exports = {
    * );
    */
   register(gulp, taskName, parameters, globals) {
-    const installDependenciesTaskName = taskName + '-install-dependencies';
+    const installDependenciesTaskName = `${taskName}-install-dependencies`;
     npmInstall.register(gulp, installDependenciesTaskName, {
       dependencies: this.dependencies
     });
 
-    const cleanUpTaskName = taskName + '-clean';
+    const cleanUpTaskName = `${taskName}-clean`;
     gulp.task(cleanUpTaskName, [installDependenciesTaskName], (done) => {
       del(['coverage'], { force: true }).then(() => {
         done();
@@ -79,7 +79,7 @@ module.exports = {
       );
     });
 
-    const watchTaskName = taskName + ':watch';
+    const watchTaskName = `${taskName}:watch`;
     gulp.task(watchTaskName, [installDependenciesTaskName], () => {
       const config = buildConfig(parameters, globals);
       gulp.watch(path.join(config.rootDir, '**/*.js'), [taskName]);