+++ src/lib/fs.js
@@ -18,9 +18,11 @@ module.exports = {
 
   readFile(filePath) {
     return new Promise((resolve, reject) => {
-      fs.readFile(filePath, 'utf8', (err, content) => {
-        return err ? reject(`File cannot be read: ${filePath}`) : resolve(content);
-      });
+      fs.readFile(
+        filePath,
+        'utf8',
+        (err, content) => err ? reject(`File cannot be read: ${filePath}`) : resolve(content)
+      );
     });
   },
 
@@ -33,9 +35,11 @@ module.exports = {
           return;
         }
 
-        fs.writeFile(targetPath, content, writeErr => {
-          return writeErr ? reject(`File cannot be written: ${targetPath}`) : resolve();
-        });
+        fs.writeFile(
+          targetPath,
+          content,
+          writeErr => writeErr ? reject(`File cannot be written: ${targetPath}`) : resolve()
+        );
       });
     });
   }